make proof system generic
